Created component TaskRow for calendar and agenda screens
